requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
httpx
python-dotenv
pydantic
pydantic-settings
loguru
pillow
python-magic; platform_system != 'Windows'
python-magic-bin; platform_system == 'Windows'
